+++ BookChat/app/src/main/java/com/example/bookchat/ui/channel/chatting/adapter/ChatItemViewHolder.kt
@@ -10,7 +10,7 @@ import com.example.bookchat.databinding.ItemChattingNoticeBinding
 import com.example.bookchat.databinding.ItemChattingOtherBinding
 import com.example.bookchat.domain.model.ChatStatus
 import com.example.bookchat.ui.channel.chatting.model.ChatItem
-import com.example.bookchat.utils.DateManager
+import com.example.bookchat.utils.getFormattedTimeText
 import com.example.bookchat.utils.image.loadUserProfile
 
 sealed class ChatItemViewHolder(
@@ -68,7 +68,7 @@ class MyChatViewHolder(
 				if (item.status == ChatStatus.LOADING || item.status == ChatStatus.RETRY_REQUIRED)
 					View.VISIBLE else View.GONE
 			chatDispatchTimeTv.text =
-				if (item.dispatchTime.isNotBlank()) DateManager.getFormattedTimeText(item.dispatchTime)
+				if (item.dispatchTime.isNotBlank()) getFormattedTimeText(item.dispatchTime)
 				else ""
 			setCaptureViewState(
 				isCaptureMode = isCaptureMode,
@@ -120,7 +120,7 @@ class AnotherUserChatViewHolder(
 				rootLayout = binding.root
 			)
 			chatDispatchTimeTv.text =
-				if (item.dispatchTime.isNotBlank()) DateManager.getFormattedTimeText(item.dispatchTime)
+				if (item.dispatchTime.isNotBlank()) getFormattedTimeText(item.dispatchTime)
 				else ""
 			userProfileIv.loadUserProfile(
 				imageUrl = item.sender?.profileImageUrl,
